Widen xcframework plist <integer> parse to 64-bit so one large value can't discard the whole Info.plist
A plist <integer> wider than int32 made the parser throw, and the reader's
surrounding try/catch then discarded the entire Info.plist — silently degrading
a third-party xcframework's version and MinimumOSVersion to a placeholder with no
diagnostic. Parse as long (the plist 64-bit integer width) and round-trip long
through the pruned-plist writer so a preserved root integer key isn't rewritten
as a string.

+++ src/Swift.Bindings/src/Configuration/XCFrameworkResolver.cs
@@ -1002,7 +1002,11 @@ namespace BindingsGeneration
                 case "string":
                     return node.InnerText;
                 case "integer":
-                    return int.Parse(node.InnerText);
+                    // plist <integer> is a 64-bit signed value (CFNumber). Parse as long so a
+                    // value wider than 32 bits (build metadata, a large UID, a timestamp) doesn't
+                    // overflow and tear down the entire dict — which would silently degrade the
+                    // version + MinimumOSVersion we actually read to a placeholder.
+                    return long.Parse(node.InnerText);
                 case "true":
                     return true;
                 case "false":
